Start converting DPD potentials.
