+++ hoomd/md/EvaluatorPairDPDThermoDPD.h
@@ -1,20 +1,20 @@
 // Copyright (c) 2009-2022 The Regents of the University of Michigan.
 // Part of HOOMD-blue, released under the BSD 3-Clause License.
 
-#ifndef __PAIR_EVALUATOR_DPDLJ_H__
-#define __PAIR_EVALUATOR_DPDLJ_H__
+#ifndef __PAIR_EVALUATOR_DPD_H__
+#define __PAIR_EVALUATOR_DPD_H__
 
 #ifndef __HIPCC__
 #include <string>
 #endif
 
 #include "hoomd/HOOMDMath.h"
+
 #include "hoomd/RNGIdentifiers.h"
 #include "hoomd/RandomNumbers.h"
 
-/*! \file EvaluatorPairDPDLJThermo.h
-    \brief Defines the pair evaluator class for the DPD Thermostat with a Lennard Jones conservative
-   potential
+/*! \file EvaluatorPairDPDThermo.h
+    \brief Defines the pair evaluator class for the DPD conservative potential
 */
 
 // need to declare these class methods with __device__ qualifiers when building in nvcc
@@ -37,32 +37,28 @@ namespace md
 
     See EvaluatorPairLJ
 
-    <b>DPD Thermostat and Conservative LJ specifics</b>
+    <b>DPD Thermostat and Conservative specifics</b>
 
-    EvaluatorPairDPDLJThermo::evalForceAndEnergy evaluates the Lennard-Jones function (see
-   EvaluatorPairLJ.  However it is not intended to be used. It is written for completeness sake
-   only.
+    EvaluatorPairDPDThermo::evalForceAndEnergy evaluates the function:
+    \f[ V_{\mathrm{DPD-C}}(r) = A \cdot \left( r_{\mathrm{cut}} - r \right)
+                        - \frac{1}{2} \cdot \frac{A}{r_{\mathrm{cut}}} \cdot
+   \left(r_{\mathrm{cut}}^2 - r^2 \right)\f]
 
-    EvaluatorPairDPDLJThermo::evalForceEnergyThermo evaluates the function:
+    The DPD Conservative potential does not need charge or diameter. One parameter is specified and
+   stored in a Scalar. \a A is placed in \a param.
+
+    EvaluatorPairDPDThermo::evalForceEnergyThermo evaluates the function:
     \f{eqnarray*}
     F =   F_{\mathrm{C}}(r) + F_{\mathrm{R,ij}}(r_{ij}) +  F_{\mathrm{D,ij}}(v_{ij}) \\
     \f}
 
     \f{eqnarray*}
-    F_{\mathrm{C}}(r) = & \partial V_{\mathrm{LJ}} / \partial r    \\
+    F_{\mathrm{C}}(r) = & A \cdot  w(r_{ij}) \\
     F_{\mathrm{R, ij}}(r_{ij}) = & - \theta_{ij}\sqrt{3} \sqrt{\frac{2k_b\gamma T}{\Delta t}}\cdot
    w(r_{ij})  \\
     F_{\mathrm{D, ij}}(r_{ij}) = & - \gamma w^2(r_{ij})\left( \hat r_{ij} \circ v_{ij} \right)  \\
     \f}
 
-    where
-     \f{eqnarray*}
-    V_{\mathrm{LJ}}(r) = & 4 \varepsilon \left[ \left( \frac{\sigma}{r} \right)^{12} -
-                                            \left( \frac{\sigma}{r} \right)^{6} \right]  & r <
-   r_{\mathrm{cut}} \\
-                                            = & 0 & r \ge r_{\mathrm{cut}} \\
-    \f}
-    and
     \f{eqnarray*}
     w(r_{ij}) = &\left( 1 - r/r_{\mathrm{cut}} \right)  & r < r_{\mathrm{cut}} \\
                      = & 0 & r \ge r_{\mathrm{cut}} \\
@@ -70,23 +66,21 @@ namespace md
     where \f$\hat r_{ij} \f$ is a normalized vector from particle i to particle j, \f$ v_{ij} = v_i
    - v_j \f$, and \f$ \theta_{ij} \f$ is a uniformly distributed random number in the range [-1, 1].
 
-    The LJ potential does not need diameter or charge. Three parameters are specified and stored in
-   a Scalar4. \a lj1 is placed in \a params.x, \a lj2 is in \a params.y and \a gamma in \a params.z.
-   The final parameter \a params.w is not used and set to zero.
+    The DPD Thermostat potential does not need charge or diameter. Two parameters are specified and
+   stored in a Scalar. \a A and \a gamma are placed in \a param.
 
-    lj1 and lj2 are related to the standard lj parameters sigma and epsilon by:
-    - \a lj1 = 4.0 * epsilon * pow(sigma,12.0)
-    - \a lj2 = 4.0 * epsilon * pow(sigma,6.0);
+    These are related to the standard lj parameters sigma and epsilon by:
+    - \a A = \f$ A \f$
+    - \a gamma = \f$ \gamma \f$
 
 */
-class EvaluatorPairDPDLJThermo
+class EvaluatorPairDPDThermoDPD
     {
     public:
     //! Define the parameter type used by this pair potential evaluator
     struct param_type
         {
-        Scalar sigma_6;
-        Scalar epsilon_x_4;
+        Scalar A;
         Scalar gamma;
 
         DEVICE void load_shared(char*& ptr, unsigned int& available_bytes) { }
@@ -94,51 +88,58 @@ class EvaluatorPairDPDLJThermo
         HOSTDEVICE void allocate_shared(char*& ptr, unsigned int& available_bytes) const { }
 
 #ifdef ENABLE_HIP
-        //! Set CUDA memory hints
-        void set_memory_hint() const
-            {
-            // default implementation does nothing
-            }
+        // CUDA memory hints
+        void set_memory_hints() const { }
 #endif
-
 #ifndef __HIPCC__
-        param_type() : sigma_6(0), epsilon_x_4(0), gamma(0) { }
+        param_type() : A(0), gamma(0) { }
 
         param_type(pybind11::dict v, bool managed = false)
             {
-            auto sigma(v["sigma"].cast<Scalar>());
-            auto epsilon(v["epsilon"].cast<Scalar>());
-            sigma_6 = sigma * sigma * sigma * sigma * sigma * sigma;
-            epsilon_x_4 = Scalar(4.0) * epsilon;
-            gamma = v["gamma"].cast<Scalar>();
-            if (gamma == 0)
-                throw std::invalid_argument(
-                    "Cannot set gamma to 0, try using DPDConservative instead.");
+            A = v["A"].cast<Scalar>();
+            // protect against a user setting gamma to 0 in dpd
+            if (v.contains("gamma"))
+                {
+                auto gam = v["gamma"].cast<Scalar>();
+                if (gam == 0)
+                    throw std::invalid_argument(
+                        "Cannot set gamma to 0 in DPD, try using DPDConservative instead.");
+                else
+                    gamma = gam;
+                }
+            else
+                gamma = 0;
             }
 
         pybind11::dict asDict()
             {
             pybind11::dict v;
-            v["sigma"] = pow(sigma_6, 1. / 6.);
-            v["epsilon"] = epsilon_x_4 / 4.0;
-            v["gamma"] = gamma;
+            v["A"] = A;
+            if (gamma != 0.0)
+                {
+                v["gamma"] = gamma;
+                }
             return v;
             }
 #endif
-        } __attribute__((aligned(16)));
+        }
+#ifdef SINGLE_PRECISION
+        __attribute__((aligned(8)));
+#else
+        __attribute__((aligned(16)));
+#endif
 
     //! Constructs the pair potential evaluator
     /*! \param _rsq Squared distance between the particles
         \param _rcutsq Squared distance at which the potential goes to 0
         \param _params Per type pair parameters of this potential
     */
-    DEVICE EvaluatorPairDPDLJThermo(Scalar _rsq, Scalar _rcutsq, const param_type& _params)
-        : rsq(_rsq), rcutsq(_rcutsq), lj1(_params.epsilon_x_4 * _params.sigma_6 * _params.sigma_6),
-          lj2(_params.epsilon_x_4 * _params.sigma_6), gamma(_params.gamma)
+    DEVICE EvaluatorPairDPDThermoDPD(Scalar _rsq, Scalar _rcutsq, const param_type& _params)
+        : rsq(_rsq), rcutsq(_rcutsq), a(_params.A), gamma(_params.gamma)
         {
         }
 
-    //! Set i and j, (particle indices, or should it be tags), and the timestep
+    //! Set i and j, (particle tags), and the timestep
     DEVICE void
     set_seed_ij_timestep(uint16_t seed, unsigned int i, unsigned int j, uint64_t timestep)
         {
@@ -166,7 +167,7 @@ class EvaluatorPairDPDLJThermo
         m_T = Temp;
         }
 
-    //! LJ does not use diameter
+    //! Does not use diameter
     DEVICE static bool needsDiameter()
         {
         return false;
@@ -177,18 +178,18 @@ class EvaluatorPairDPDLJThermo
     */
     DEVICE void setDiameter(Scalar di, Scalar dj) { }
 
-    //! LJ doesn't use charge
+    //! Yukawa doesn't use charge
     DEVICE static bool needsCharge()
         {
         return false;
         }
-    //! Accept the optional charge values
+    //! Accept the optional diameter values
     /*! \param qi Charge of particle i
         \param qj Charge of particle j
     */
     DEVICE void setCharge(Scalar qi, Scalar qj) { }
 
-    //! Evaluate the force and energy
+    //! Evaluate the force and energy using the conservative force only
     /*! \param force_divr Output parameter to write the computed force divided by r.
         \param pair_eng Output parameter to write the computed pair energy
         \param energy_shift If true, the potential must be shifted so that V(r) is continuous at the
@@ -200,20 +201,17 @@ class EvaluatorPairDPDLJThermo
     DEVICE bool evalForceAndEnergy(Scalar& force_divr, Scalar& pair_eng, bool energy_shift)
         {
         // compute the force divided by r in force_divr
-        if (rsq < rcutsq && lj1 != 0)
+        if (rsq < rcutsq)
             {
-            Scalar r2inv = Scalar(1.0) / rsq;
-            Scalar r6inv = r2inv * r2inv * r2inv;
-            force_divr = r2inv * r6inv * (Scalar(12.0) * lj1 * r6inv - Scalar(6.0) * lj2);
+            Scalar rinv = fast::rsqrt(rsq);
+            Scalar r = Scalar(1.0) / rinv;
+            Scalar rcutinv = fast::rsqrt(rcutsq);
+            Scalar rcut = Scalar(1.0) / rcutinv;
 
-            pair_eng = r6inv * (lj1 * r6inv - lj2);
+            // force is easy to calculate
+            force_divr = a * (rinv - rcutinv);
+            pair_eng = a * (rcut - r) - Scalar(1.0 / 2.0) * a * rcutinv * (rcutsq - rsq);
 
-            if (energy_shift)
-                {
-                Scalar rcut2inv = Scalar(1.0) / rcutsq;
-                Scalar rcut6inv = rcut2inv * rcut2inv * rcut2inv;
-                pair_eng -= rcut6inv * (lj1 * rcut6inv - lj2);
-                }
             return true;
             }
         else
@@ -221,30 +219,29 @@ class EvaluatorPairDPDLJThermo
         }
 
     //! Evaluate the force and energy using the thermostat
-    /*! \param force_divr Output parameter to write the computed total force divided by r.
+    /*! \param force_divr Output parameter to write the computed force divided by r.
         \param force_divr_cons Output parameter to write the computed conservative force divided by
-       r. \param pair_eng Output parameter to write the computed pair energy \param energy_shift If
-       true, the potential must be shifted so that V(r) is continuous at the cutoff \note There is
-       no need to check if rsq < rcutsq in this method. Cutoff tests are performed in PotentialPair.
+       r. \param pair_eng Output parameter to write the computed pair energy \param energy_shift
+       Ignored. DPD always goes to 0 at the cutoff. \note There is no need to check if rsq < rcutsq
+       in this method. Cutoff tests are performed in PotentialPair.
 
         \note The conservative part \b only must be output to \a force_divr_cons so that the virial
        may be computed correctly.
 
         \return True if they are evaluated or false if they are not because we are beyond the cutoff
     */
-
     DEVICE bool evalForceEnergyThermo(Scalar& force_divr,
                                       Scalar& force_divr_cons,
                                       Scalar& pair_eng,
                                       bool energy_shift)
         {
         // compute the force divided by r in force_divr
-        if (rsq < rcutsq && lj1 != 0)
+        if (rsq < rcutsq)
             {
             Scalar rinv = fast::rsqrt(rsq);
-            Scalar r2inv = Scalar(1.0) / rsq;
-            Scalar r6inv = r2inv * r2inv * r2inv;
+            Scalar r = Scalar(1.0) / rinv;
             Scalar rcutinv = fast::rsqrt(rcutsq);
+            Scalar rcut = Scalar(1.0) / rcutinv;
 
             // force calculation
 
@@ -268,8 +265,11 @@ class EvaluatorPairDPDLJThermo
             // Generate a single random number
             Scalar alpha = hoomd::UniformDistribution<Scalar>(-1, 1)(rng);
 
-            // conservative lj
-            force_divr = r2inv * r6inv * (Scalar(12.0) * lj1 * r6inv - Scalar(6.0) * lj2);
+            // conservative dpd
+            // force_divr = FDIV(a,r)*(Scalar(1.0) - r*rcutinv);
+            force_divr = a * (rinv - rcutinv);
+
+            //  conservative force only
             force_divr_cons = force_divr;
 
             //  Drag Term
@@ -280,14 +280,7 @@ class EvaluatorPairDPDLJThermo
                 += fast::rsqrt(m_deltaT / (m_T * gamma * Scalar(6.0))) * (rinv - rcutinv) * alpha;
 
             // conservative energy only
-            pair_eng = r6inv * (lj1 * r6inv - lj2);
-
-            if (energy_shift)
-                {
-                Scalar rcut2inv = Scalar(1.0) / rcutsq;
-                Scalar rcut6inv = rcut2inv * rcut2inv * rcut2inv;
-                pair_eng -= rcut6inv * (lj1 * rcut6inv - lj2);
-                }
+            pair_eng = a * (rcut - r) - Scalar(1.0 / 2.0) * a * rcutinv * (rcutsq - rsq);
 
             return true;
             }
@@ -311,7 +304,7 @@ class EvaluatorPairDPDLJThermo
      */
     static std::string getName()
         {
-        return std::string("dpdlj");
+        return std::string("dpd");
         }
 
     std::string getShapeSpec() const
@@ -323,8 +316,7 @@ class EvaluatorPairDPDLJThermo
     protected:
     Scalar rsq;          //!< Stored rsq from the constructor
     Scalar rcutsq;       //!< Stored rcutsq from the constructor
-    Scalar lj1;          //!< lj1 parameter extracted from the params passed to the constructor
-    Scalar lj2;          //!< lj2 parameter extracted from the params passed to the constructor
+    Scalar a;            //!< a parameter for potential extracted from params by constructor
     Scalar gamma;        //!< gamma parameter for potential extracted from params by constructor
     uint16_t m_seed;     //!< User set seed for thermostat PRNG
     unsigned int m_i;    //!< index of first particle (should it be tag?).  For use in PRNG
@@ -340,4 +332,4 @@ class EvaluatorPairDPDLJThermo
     } // end namespace md
     } // end namespace hoomd
 
-#endif // __PAIR_EVALUATOR_DPDLJ_H__
+#endif // __PAIR_EVALUATOR_DPD_H__
