+++ hoomd/md/PotentialPairDPDThermoGPU.h
@@ -40,9 +40,7 @@ namespace md
 
     \sa export_PotentialPairDPDThermoGPU()
 */
-template<class evaluator,
-         hipError_t gpu_cpdf(const kernel::dpd_pair_args_t& pair_args,
-                             const typename evaluator::param_type* d_params)>
+template<class evaluator>
 class PotentialPairDPDThermoGPU : public PotentialPairDPDThermo<evaluator>
     {
     public:
@@ -82,10 +80,8 @@ class PotentialPairDPDThermoGPU : public PotentialPairDPDThermo<evaluator>
     virtual void computeForces(uint64_t timestep);
     };
 
-template<class evaluator,
-         hipError_t gpu_cpdf(const kernel::dpd_pair_args_t& pair_args,
-                             const typename evaluator::param_type* d_params)>
-PotentialPairDPDThermoGPU<evaluator, gpu_cpdf>::PotentialPairDPDThermoGPU(
+template<class evaluator>
+PotentialPairDPDThermoGPU<evaluator>::PotentialPairDPDThermoGPU(
     std::shared_ptr<SystemDefinition> sysdef,
     std::shared_ptr<NeighborList> nlist)
     : PotentialPairDPDThermo<evaluator>(sysdef, nlist), m_param(0)
@@ -120,10 +116,8 @@ PotentialPairDPDThermoGPU<evaluator, gpu_cpdf>::PotentialPairDPDThermoGPU(
 #endif
     }
 
-template<class evaluator,
-         hipError_t gpu_cpdf(const kernel::dpd_pair_args_t& pair_args,
-                             const typename evaluator::param_type* d_params)>
-void PotentialPairDPDThermoGPU<evaluator, gpu_cpdf>::computeForces(uint64_t timestep)
+template<class evaluator>
+void PotentialPairDPDThermoGPU<evaluator>::computeForces(uint64_t timestep)
     {
     this->m_nlist->compute(timestep);
 
@@ -175,7 +169,7 @@ void PotentialPairDPDThermoGPU<evaluator, gpu_cpdf>::computeForces(uint64_t time
     unsigned int block_size = param / 10000;
     unsigned int threads_per_particle = param % 10000;
 
-    gpu_cpdf(kernel::dpd_pair_args_t(d_force.data,
+    kernel::gpu_compute_dpd_forces<evaluator>(kernel::dpd_pair_args_t(d_force.data,
                                      d_virial.data,
                                      this->m_virial.getPitch(),
                                      this->m_pdata->getN(),
@@ -210,16 +204,15 @@ namespace detail
     {
 //! Export this pair potential to python
 /*! \param name Name of the class in the exported python module
-    \tparam T Class type to export. \b Must be an instantiated PotentialPairDPDThermoGPU class
-   template. \tparam Base Base class of \a T. \b Must be PotentialPairDPDThermo<evaluator> with the
-   same evaluator as used in \a T.
+    \tparam T Evaluator type to export.
+
 */
-template<class T, class Base>
+template<class T>
 void export_PotentialPairDPDThermoGPU(pybind11::module& m, const std::string& name)
     {
-    pybind11::class_<T, Base, std::shared_ptr<T>>(m, name.c_str())
+    pybind11::class_<PotentialPairDPDThermoGPU<T>, PotentialPairDPDThermo<T>, std::shared_ptr<PotentialPairDPDThermoGPU<T>>>(m, name.c_str())
         .def(pybind11::init<std::shared_ptr<SystemDefinition>, std::shared_ptr<NeighborList>>())
-        .def("setTuningParam", &T::setTuningParam);
+        .def("setTuningParam", &PotentialPairDPDThermoGPU<T>::setTuningParam);
     }
 
     } // end namespace detail
